refactor: move more items to courseAPI.ts ...
roughly:
- courseDB should "only" be the class-based interface to courses.
- courseAPI is the direct functional methods
