+++ packages/vue/src/db/courseAPI.ts
@@ -1,9 +1,14 @@
+import pouch from 'pouchdb-browser';
+import { pouchDBincludeCredentialsConfig } from '.';
+import ENV from '../ENVIRONMENT_VARS';
 import { DataShape } from '../base-course/Interfaces/DataShape';
 import { FieldDefinition } from '../base-course/Interfaces/FieldDefinition';
+import { NameSpacer, ShapeDescriptor } from '../courses/NameSpacer';
 import { FieldType } from '../enums/FieldType';
-import { DisplayableData, DocType } from './types';
-import { getCourseDB, createCards } from './courseDB';
-import { NameSpacer } from '../courses/NameSpacer';
+import { CourseConfig } from '../server/types';
+import { CourseElo, blankCourseElo, toCourseElo } from '../tutor/Elo';
+import { CourseDB, createTag, updateCardElo } from './courseDB';
+import { CardData, DisplayableData, DocType, Tag } from './types';
 
 /**
  *
@@ -119,4 +124,149 @@ export async function addNote55(
   }
 
   return result;
+}
+
+export async function createCards(
+  courseID: string,
+  datashapeID: PouchDB.Core.DocumentId,
+  noteID: PouchDB.Core.DocumentId,
+  tags: string[]
+) {
+  const cfg = await getCredentialledCourseConfig(courseID);
+  const dsDescriptor = NameSpacer.getDataShapeDescriptor(datashapeID);
+  let questionViewTypes: string[] = [];
+
+  for (const ds of cfg.dataShapes) {
+    if (ds.name === datashapeID) {
+      questionViewTypes = ds.questionTypes;
+    }
+  }
+
+  for (const questionView of questionViewTypes) {
+    createCard(questionView, courseID, dsDescriptor, noteID, tags);
+  }
+}
+
+async function createCard(
+  questionViewName: string,
+  courseID: string,
+  dsDescriptor: ShapeDescriptor,
+  noteID: string,
+  tags: string[]
+) {
+  tags.forEach((t) => console.log(`Adding ${t}!`));
+  const qDescriptor = NameSpacer.getQuestionDescriptor(questionViewName);
+  const cfg = await getCredentialledCourseConfig(courseID);
+
+  for (const rQ of cfg.questionTypes) {
+    if (rQ.name === questionViewName) {
+      for (const view of rQ.viewList) {
+        addCard(
+          courseID,
+          dsDescriptor.course,
+          [noteID],
+          NameSpacer.getViewString({
+            course: qDescriptor.course,
+            questionType: qDescriptor.questionType,
+            view,
+          }),
+          blankCourseElo(),
+          tags
+        );
+      }
+    }
+  }
+}
+
+/**
+ * Adds a card to the DB. This function is called
+ * as a side effect of adding either a View or
+ * DisplayableData item.
+ * @param course The name of the course that the card belongs to
+ * @param id_displayable_data C/PouchDB ID of the data used to hydrate the view
+ * @param id_view C/PouchDB ID of the view used to display the card
+ */
+async function addCard(
+  courseID: string,
+  course: string,
+  id_displayable_data: PouchDB.Core.DocumentId[],
+  id_view: PouchDB.Core.DocumentId,
+  elo: CourseElo,
+  tags: string[]
+) {
+  const card = await getCourseDB(courseID).post<CardData>({
+    course,
+    id_displayable_data,
+    id_view,
+    docType: DocType.CARD,
+    elo: elo || toCourseElo(990 + Math.round(20 * Math.random())),
+  });
+  tags.forEach((tag) => {
+    console.log(`adding tag: ${tag} to card ${card.id}`);
+    addTagToCard(courseID, card.id, tag);
+  });
+  return card;
+}
+
+export async function getCredentialledCourseConfig(courseID: string) {
+  const db = getCourseDB(courseID);
+  const ret = await db.get<CourseConfig>('CourseConfig');
+  ret.courseID = courseID;
+  console.log(`Returning corseconfig:
+
+  ${JSON.stringify(ret)}
+  `);
+
+  return ret;
+}
+
+export async function addTagToCard(
+  courseID: string,
+  cardID: string,
+  tagID: string
+): Promise<PouchDB.Core.Response> {
+  // todo: possible future perf. hit if tags have large #s of taggedCards.
+  // In this case, should be converted to a server-request
+  const prefixedTagID = getTagID(tagID);
+  const courseDB = getCourseDB(courseID);
+  const courseApi = new CourseDB(courseID);
+  try {
+    console.log(`Applying tag ${tagID} to card ${courseID + '-' + cardID}...`);
+    const tag = await courseDB.get<Tag>(prefixedTagID);
+    if (!tag.taggedCards.includes(cardID)) {
+      tag.taggedCards.push(cardID);
+
+      courseApi.getCardEloData([cardID]).then((eloData) => {
+        const elo = eloData[0];
+        elo.tags[tagID] = {
+          count: 0,
+          score: elo.global.score, // todo: or 1000?
+        };
+        updateCardElo(courseID, cardID, elo);
+      });
+
+      return courseDB.put<Tag>(tag);
+    } else throw new Error(`Card already has this tag`);
+  } catch (e) {
+    console.log(`Tag ${tagID} does not exist...`);
+    await createTag(courseID, tagID);
+    return addTagToCard(courseID, cardID, tagID);
+  }
+}
+
+export function getTagID(tagName: string): string {
+  const tagPrefix = DocType.TAG.valueOf() + '-';
+  if (tagName.indexOf(tagPrefix) === 0) {
+    return tagName;
+  } else {
+    return tagPrefix + tagName;
+  }
+}
+
+export function getCourseDB(courseID: string): PouchDB.Database {
+  const dbName = `coursedb-${courseID}`;
+  return new pouch(
+    ENV.COUCHDB_SERVER_PROTOCOL + '://' + ENV.COUCHDB_SERVER_URL + dbName,
+    pouchDBincludeCredentialsConfig
+  );
 }
